requirements.txt
appdirs==1.4.4
appnope==0.1.2
astroid==2.9.0
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.9.3
black==21.7b0
bs4==0.0.1
CacheControl==0.12.10
cachy==0.3.0
certifi==2021.5.30
charset-normalizer==2.0.4
cleo==0.8.1
click==8.0.1
clikit==0.6.2
crashtest==0.3.1
debugpy==1.4.1
decorator==5.0.9
distlib==0.3.3
entrypoints==0.3
fontawesomefree==5.15.4
filelock==3.4.0
Flask-Cors==3.0.10
Flask==2.0.2
glob2==0.7
html5lib==1.1
idna==3.2
iniconfig==1.1.1
ipykernel==6.2.0
ipython-genutils==0.2.0
ipython==7.26.0
isort==5.10.1
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.3
jupyter-client==7.0.1
jupyter-core==4.7.1
keyring==21.8.0
lazy-object-proxy==1.6.0
lockfile==0.12.2
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
msgpack==1.0.2
mypy-extensions==0.4.3
nest-asyncio==1.5.1
numpy==1.21.2
packaging==20.9
parse-type==0.5.2
parse==1.19.0
parso==0.8.2
pastel==0.2.1
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
pip==21.3.1
pipenv==2021.11.15
pkginfo==1.8.1
platformdirs==2.4.0
pluggy==0.13.1
poetry-core==1.0.7
poetry==1.1.11
prompt-toolkit==3.0.20
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.7.0
pycryptodome==3.11.0
Pygments==2.10.0
pylev==1.4.0
pylint==2.12.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.2
pytest-bdd==4.1.0
pytest==6.2.5
python-dateutil==2.8.2
pyzmq==22.2.1
regex==2021.8.3
requests-toolbelt==0.9.1
requests==2.26.0
selenium==3.141.0
setuptools==49.2.1
shellingham==1.4.0
six==1.15.0
soupsieve==2.2.1
toml==0.10.2
tomli==1.2.1
tomlkit==0.7.2
tornado==6.1
traitlets==5.0.5
typing-extensions==4.0.0
urllib3==1.26.6
virtualenv-clone==0.5.7
virtualenv==20.10.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wheel==0.33.1
wrapt==1.13.3
yapf==0.31.0
